+++ Bencode/Types/Int.cpp
@@ -4,7 +4,7 @@ bencode::Int::Int()
 	:Int(0)
 {}
 
-bencode::Int::Int(const size_t& value)
+bencode::Int::Int(const uint64_t& value)
 	: mValue(value)
 {}
 
@@ -17,7 +17,7 @@ bencode::Int::Int(Int&& other) noexcept
 {
 }
 
-size_t bencode::Int::getValue() const
+uint64_t bencode::Int::getValue() const
 {
 	return mValue;
 }
@@ -33,7 +33,7 @@ std::string bencode::Int::code() const
 	return result;
 }
 
-bencode::Int& bencode::Int::operator=(const size_t& value)
+bencode::Int& bencode::Int::operator=(const uint64_t& value)
 {
 	mValue = value;
 
@@ -47,7 +47,7 @@ bencode::Int& bencode::Int::operator=(const Int& other)
 	return *this;
 }
 
-bencode::Int bencode::Int::operator+(const size_t& b) const
+bencode::Int bencode::Int::operator+(const uint64_t& b) const
 {
 	return Int(mValue + b);
 }
@@ -57,7 +57,7 @@ bencode::Int bencode::Int::operator+(const Int& b) const
 	return Int(mValue + b.mValue);
 }
 
-bencode::Int bencode::Int::operator-(const size_t& b) const
+bencode::Int bencode::Int::operator-(const uint64_t& b) const
 {
 	return Int(mValue - b);
 }
@@ -67,12 +67,12 @@ bencode::Int bencode::Int::operator-(const Int& b) const
 	return Int(mValue - b.mValue);
 }
 
-bencode::Int operator+(const size_t& a, const bencode::Int& b)
+bencode::Int operator+(const uint64_t& a, const bencode::Int& b)
 {
 	return b + a;
 }
 
-bencode::Int operator-(const size_t& a, const bencode::Int& b)
+bencode::Int operator-(const uint64_t& a, const bencode::Int& b)
 {
 	return b - a;
 }
