+++ Client/File.h
@@ -17,20 +17,20 @@ class File
 	QMutex mutex;
 	QFile* mFile;
 	QVector <FileFrag*> mFileFragments;
-	size_t mFragSize;
-	size_t mFileSize;
+	uint64_t mFragSize;
+	uint64_t mFileSize;
 	bool mIsReadonly = true;
 public:
-	File(const QString& name, const size_t& fragSize,  QObject *parent = nullptr);
-	File(const QString& name, const size_t& fragSize, const size_t& fileSize, QObject* parent = nullptr);
+	File(const QString& name, const uint64_t& fragSize,  QObject *parent = nullptr);
+	File(const QString& name, const uint64_t& fragSize, const uint64_t& fileSize, QObject* parent = nullptr);
 	~File();
 
 	FileFrag* operator [] (const unsigned int index);
 
 	bool readonly() const;
 	void readonly(bool isReadonly);
-	const size_t getSize() const;
-	const size_t getFragSize() const;
+	const uint64_t getSize() const;
+	const uint64_t getFragSize() const;
 	const unsigned int getFragCount() const;
 	const std::string getFragsHash() const;
